make changes in structure with make verision in controller, requests and resources
