+++ app/Domains/Roles/V1/Requests/AddRequest.php
@@ -1,0 +1,45 @@
+<?php
+
+namespace App\Domains\Roles\V1\Requests;
+
+use App\Http\Traits\GeneralTrait;
+use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Contracts\Validation\Validator;
+use Illuminate\Http\Exceptions\HttpResponseException;
+
+class AddRequest extends FormRequest
+{
+    use GeneralTrait;
+    /**
+     * Determine if the user is authorized to make this request.
+     */
+    public function authorize(): bool
+    {
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
+     */
+    public function rules(): array
+    {
+        return [
+            //
+        ];
+    }
+    /**
+     * Handle a failed validation attempt.
+     *
+     * @param  \Illuminate\Contracts\Validation\Validator  $validator
+     * @return void
+     *
+     * @throws \Illuminate\Http\Exceptions\HttpResponseException
+     */
+    protected function failedValidation(Validator $validator): void
+    {
+        $code = $this->returnCodeAccordingToInput($validator);
+        throw new HttpResponseException($this->returnValidationError($code, $validator));
+    }
+}
