--- app/Domains/V1/Roles/Requests/AddRequest.php
@@ -1,28 +1,0 @@
-<?php
-
-namespace App\Domains\V1\Roles\Requests;
-
-use Illuminate\Foundation\Http\FormRequest;
-
-class AddRequest extends FormRequest
-{
-    /**
-     * Determine if the user is authorized to make this request.
-     */
-    public function authorize(): bool
-    {
-        return true;
-    }
-
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
-     */
-    public function rules(): array
-    {
-        return [
-            //
-        ];
-    }
-}
